working many to many table thingy

+++ app/Http/Controllers/TrailsController.php
@@ -7,6 +7,8 @@ use App\Stop;
 use Illuminate\Http\Request;
 use App\Trail;
 use Image;
+use File;
+
 class TrailsController extends Controller
 {
     /**
@@ -31,82 +33,74 @@ class TrailsController extends Controller
     public function create()
     {
         $stops = Stop::all();
-         return view('trails.create', compact('stops'));
+        return view('trails.create', compact('stops'));
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
-     
-      $validated = request()->validate([
+
+        $validated = request()->validate([
             'name' => ['required'],
             'length' => ['required'],
             'time' => ['required'],
 
-       ]);
+        ]);
 
-      Trail::create($validated);
+        $trail = Trail::create($validated);
 
-
-      $trail = Trail::where('name', $request->get('name'))->get();
-//      $trail->stops()->attach($request->get('stops'));
-//        dd($trail);
-        $trail->stops()->get()->attach(1);
-//        $stops=$request->get('stops');
-//        dd($stops);
-//        foreach ($stops as $stop)
-//        {
-////          dd((int)$stop);
-//            $trail->stops()->sync((int)$stop);
-//        }
+        $stops = $request->get('stops');
+        foreach ($stops as $stop) {
+            $trail->stops()->attach((int)$stop);
+        }
 
 
-     $trail = Trail::create($validated);
-        if($request->hasFile('img')) {
+        if ($request->hasFile('img')) {
             $img = $request->file('img');
             $filename = time() . '.' . $img->getClientOriginalExtension();
             Image::make($img)->resize(300, 300)->save(public_path('/images/trails/' . $filename));
             $trail->img = $filename;
             $trail->save();
         }
-         return redirect('/trails');
+        return redirect('/trails');
 
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function show($id)
     {
         $trail = Trail::findOrFail($id);
-     return view('trails.show', compact('trail'));
+        return view('trails.show', compact('trail'));
     }
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function edit($id)
     {
+        $stops = Stop::all();
         $trail = Trail::findOrFail($id);
-        return view('trails.edit', compact('trail'));
+        return view('trails.edit', compact('trail', 'stops'));
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param \Illuminate\Http\Request $request
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $id)
@@ -116,47 +110,63 @@ class TrailsController extends Controller
             'length' => ['required'],
             'time' => ['required'],
         ]);
+
         $trail = trail::findOrFail($id);
         $trail->name = $request->get('name');
         $trail->length = $request->get('length');
         $trail->time = $request->get('time');
 
-        if($request->hasFile('img')) {
-            if($trail->img == 'default.jpg') {
+        if ($request->hasFile('img')) {
+            if ($trail->img == 'default.jpg') {
                 $img = $request->file('img');
                 $filename = time() . '.' . $img->getClientOriginalExtension();
-                Image::make($img)->resize(300, 300)->save(public_path('/images/stops/' . $filename));
+                Image::make($img)->resize(300, 300)->save(public_path('/images/trails/' . $filename));
                 $trail->img = $filename;
-            }else{
-                $image = public_path('/images/stops/' .  $trail->img);
+            } else {
+                $image = public_path('/images/trails/' . $trail->img);
                 File::delete($image);
                 $img = $request->file('img');
                 $filename = time() . '.' . $img->getClientOriginalExtension();
-                Image::make($img)->resize(300, 300)->save(public_path('/images/stops/' . $filename));
+                Image::make($img)->resize(300, 300)->save(public_path('/images/trails/' . $filename));
                 $trail->img = $filename;
             }
-        $trail->save();
-
-        return redirect('/trails');
         }
+
+        foreach ($trail->stops as $stop) {
+            $trail->stops()->detach($stop->id);
+        }
+        $stops = $request->get('stops');
+        foreach ($stops as $stop) {
+            $trail->stops()->attach((int)$stop);
+        }
+        
+        $trail->save();
+        return redirect('/trails');
+
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param int $id
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
         $trail = Trail::findOrFail($id);
-        $trail->stops()->detach($id);
+        foreach ($trail->stops as $stop) {
+            $trail->stops()->detach($stop->id);
+        }
+
+
+        if ($trail->img != 'default.jpg') {
+            $image = public_path('/images/stops/' . $trail->img);
+            File::delete($image);
+        }
+
         $trail->delete();
-        $image = public_path('/images/stops/' .  $trail->img);
-        File::delete($image);
         return redirect('/trails');
     }
-
 
 
     public function apiAll()
